fix : 카테고리 컨텐츠 저장 안되는 문제수정
