+++ flaschenbook-spring-app/flaschenbook/src/main/java/com/book/flaschenbook/repository/SurveyContentRepository.java
@@ -1,6 +1,7 @@
 package com.book.flaschenbook.repository;
 
 import com.book.flaschenbook.entity.SurveyContentEntity;
+import com.book.flaschenbook.entity.SurveyContentId;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
@@ -9,7 +10,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface SurveyContentRepository extends JpaRepository<SurveyContentEntity, Long> {
+public interface SurveyContentRepository extends JpaRepository<SurveyContentEntity, SurveyContentId> {
     List<SurveyContentEntity> findByType(String type);
 
 }
